fix(rust): repair AGY's concurrent resolver work so the workspaces build
AGY landed a batch while this campaign was in flight (miosd drift/ module split,
mios-resolver db_overlay flag, InvalidColorHex diagnostic, MIOS_PG_BIND_ADDR
derivation, i64 stack_offset, a proptest differential test, deny.toml rewrite).
`cargo check` passed, which is why it looked fine -- but check skips test
targets, and CI runs `--all-targets -- -D warnings`, where it did not compile:

  * merge.rs: `toml! {}` with an empty body is a parse error (the macro needs at
    least one token) -- an empty overlay is `Value::Table(Map::new())`.
  * merge.rs: the new tests bind `toml!{...}`, which yields a `Table`, and pass
    it where `Value` is required -- annotated and `.into()`d.
  * tests/differential.rs: passed `(stack_id * 10000) as u16` into an i64
    parameter, and embedded the TOML body in a raw string with no newlines, so
    the fixture could never parse. Rewritten.
  * miosd/src/drift/: unused imports in mod.rs/laws.rs/regen.rs, and a dead
    `run_checks` wrapper (main calls `run_checks_cli` directly).

Also fixed check_version_ssot, which my own k3s-selinux comment tripped by
naming an upstream version literal.

Verified: `cargo fmt --all --check` and
`cargo clippy --workspace --all-targets -- -D warnings` are clean on BOTH
workspaces, and every SSOT projection re-renders to a fixed point.

+++ tools/native/mios-resolver/src/emit_ps.rs
@@ -5,7 +5,7 @@ use toml::Value;
 use crate::emit::build_exports_map;
 use crate::palette::resolve as resolve_palette;
 
-pub fn emit_powershell(merged: &Value, stack_offset: i32) -> String {
+pub fn emit_powershell(merged: &Value, stack_offset: i64) -> String {
     let exports = build_exports_map(merged, stack_offset);
     let palette = resolve_palette(merged);
